+++ lib/model/settings/defaults.dart
@@ -2,6 +2,7 @@ import 'package:aves/model/actions/entry_actions.dart';
 import 'package:aves/model/actions/entry_set_actions.dart';
 import 'package:aves/model/filters/favourite.dart';
 import 'package:aves/model/filters/mime.dart';
+import 'package:aves/model/filters/recent.dart';
 import 'package:aves/model/naming_pattern.dart';
 import 'package:aves/model/settings/enums/enums.dart';
 import 'package:aves/model/source/enums.dart';
@@ -41,6 +42,7 @@ class SettingsDefaults {
     null,
     MimeFilter.video,
     FavouriteFilter.instance,
+    RecentlyAddedFilter.instance,
   ];
   static const drawerPageBookmarks = [
     AlbumListPage.routeName,