+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 - Collection / Info: edit description via Exif / IPTC / XMP
 - Info: read XMP from HEIC on Android >=11
 - Collection: support HEIC motion photos on Android >=11
+- Search: `recently added` filter
 - Dutch translation (thanks Martijn Fabrie, Koen Koppens)
 
 ### Changed